Fixed circular imports, moved function definitions out of init file
