+++ xpsi/Data.py
@@ -2,7 +2,7 @@
 __all__ = ["Data"]
 
 from xpsi.global_imports import *
-from xpsi.__init__ import make_verbose
+from xpsi.utils import make_verbose
 
 from xpsi.Instrument import ChannelError
