+++ xpsi/Instrument.py
@@ -2,7 +2,7 @@ __all__ = ["Instrument"]
 
 from xpsi.global_imports import *
 
-from xpsi.__init__ import make_verbose
+from xpsi.utils import make_verbose
 
 from xpsi.ParameterSubspace import ParameterSubspace
