+++ xpsi/Prior.py
@@ -4,7 +4,7 @@ from xpsi import _comm, _size, _rank # of MPI.COMM_WORLD
 
 from xpsi.global_imports import *
 
-from xpsi import make_verbose
+from xpsi.utils import make_verbose
 
 from abc import ABCMeta, abstractmethod
 from xpsi.ParameterSubspace import ParameterSubspace
