+++ xpsi/Star.py
@@ -1,4 +1,3 @@
-#from .global_imports import *
 
 from xpsi.ParameterSubspace import ParameterSubspace
